update codes, format AUTO codes
